+++ src/riggedPoly.cpp
@@ -216,7 +216,7 @@ void Skeleton::init(vector<complex<double>>* polyAsset, vector<jointInfo>* joint
         headIndex = indexQueue.front();
         indexQueue.pop();
 
-        for(int j = 0; j < (*linkingInfo)[headIndex].indices.size(); j++){
+        for(long unsigned int j = 0; j < (*linkingInfo)[headIndex].indices.size(); j++){
             tailIndex = (*linkingInfo)[headIndex].indices[j];
     
             indexQueue.push(tailIndex);
